feat: Enable mentions and notify users from any text field
