+++ frappe/public/js/frappe/form/controls/text_editor.js
@@ -174,9 +174,33 @@ frappe.ui.form.ControlTextEditor = class ControlTextEditor extends frappe.ui.for
 				toolbar: this.get_toolbar_options(),
 				table: true,
 				imageResize: {},
-				magicUrl: true
+				magicUrl: true,
+				mention: this.get_mention_options()
 			},
-			theme: 'snow'
+			theme: 'snow',
+		};
+	}
+
+	get_mention_options() {
+		if (!this.enable_mentions && !this.df.enable_mentions) {
+			return null;
+		}
+		let me = this;
+		return {
+			allowedChars: /^[A-Za-z0-9_]*$/,
+			mentionDenotationChars: ["@"],
+			isolateCharacter: true,
+			source: frappe.utils.debounce(async function(search_term, renderList) {
+				let method = me.mention_search_method || 'frappe.desk.search.get_names_for_mentions';
+				let values = await frappe.xcall(method, {
+					search_term
+				});
+				renderList(values, search_term);
+			}, 300),
+			renderItem(item) {
+				let value = item.value;
+				return `${value} ${item.is_group ? frappe.utils.icon('users') : ''}`;
+			}
 		};
 	}
